payroll-provider/requirements.txt
fastapi[standard]
strawberry-graphql[fastapi]